Divided shortform helper for repository and models
