+++ src/pocketdb/models/shortform/ShortForm.cpp
@@ -2,7 +2,7 @@
 // Distributed under the Apache 2.0 software license, see the accompanying
 // https://www.apache.org/licenses/LICENSE-2.0
 
-#include "pocketdb/helpers/ShortFormHelper.h"
+#include "pocketdb/helpers/ShortFormModelsHelper.h"
 
 PocketDb::ShortForm::ShortForm(PocketDb::ShortTxType type, ShortTxData txData, std::optional<ShortTxData> relatedContent)
     : m_type(type),
